+++ src/nvim/match.h
@@ -1,0 +1,12 @@
+#ifndef NVIM_MATCH_H
+#define NVIM_MATCH_H
+
+#include "nvim/buffer_defs.h"
+#include "nvim/eval/funcs.h"
+#include "nvim/ex_cmds_defs.h"
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "match.h.generated.h"
+#endif
+
+#endif  // NVIM_MATCH_H
